requirements-test.txt
pytest-aiohttp==0.3.0
pytest-env==0.6.2
redis==3.4.1  # for testing purpose
